requirements.txt
babel==2.17.0
blinker==1.8.2
click==8.1.8
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.20.0
flask==3.0.3
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.6
lxml==6.0.0
MarkupSafe==2.1.5
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
six==1.17.0
tqdm==4.67.1
typing-extensions==4.13.2
tzdata==2025.2
werkzeug==3.0.6
zipp==3.20.2
